third_party/WebKit/LayoutTests/animations/timing/implicit-keyframe-with-timing-function.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="target">Test target</div>
<script>
var target = document.getElementById('target');
var test1 = async_test('Timing function should apply even when keyframes are not specified');
target.addEventListener('animationstart', function() {
  test1.step(function() {
    assert_equals(getComputedStyle(target).left, '1000px');
  });
  target.remove();
  test1.done();
});
</script>
<style>
#target {
  animation: anim 1s -0.25s step-start paused;
  left: 0px;
}
@keyframes anim {
  50% { left: 1000px; }
}
</style>
